==> 1_get_links.ipynb <==
import csv
import time
import requests
from bs4 import BeautifulSoup

base_url = "https://krisha.kz"
search_url = "https://krisha.kz/prodazha/kvartiry/almaty/"

# Initialize a list to store apartment links
apartment_links = []

# in range() you need to change from what page and to what page
for page_num in range(1, 2):
    # Append the current page number to the search URL
    current_page_url = f"{search_url}?page={page_num}"

    # Send request and parse the page
    response = requests.get(current_page_url)
    time.sleep(1)
    soup = BeautifulSoup(response.text, 'html.parser')

    # Check if the page has any listings
    listings = soup.find_all("a", class_="a-card__title")
    # print(listings)
    if not listings:
        break  # No more listings, exit the loop

    # Extract and store the links
    for link in listings:
        apartment_links.append(base_url + link.get('href'))

# Print the total number of links found
print(f"Found {len(apartment_links)} apartment links.")

with open('links.csv', 'a', newline='', encoding='utf-8') as csvfile:
    csv_writer = csv.writer(csvfile)
    for apartment_link in apartment_links:
        csv_writer.writerow({apartment_link})
