==> 4_combain_datas.ipynb <==
import pandas as pd

files = ['apartments_data_00000-01000.csv', 'apartments_data_01000-02000.csv', 
         'apartments_data_02000-03000.csv', 'apartments_data_03000-04000.csv', 
         'apartments_data_04000-06000.csv', 'apartments_data_06000-08000.csv', 
         'apartments_data_08000-10000.csv', 
         'apartments_data_10000-11000.csv', 'apartments_data_11000-12000.csv', 
         'apartments_data_12000-13000.csv', 'apartments_data_13000-14000.csv', 
         'apartments_data_14000-15000.csv', 'apartments_data_15000-16000.csv', 
         'apartments_data_16000-17000.csv', 'apartments_data_17000-18000.csv', 
         'apartments_data_18000-19000.csv', 'apartments_data_19000-20000.csv', 
         'apartments_data_20000-21000.csv', 'apartments_data_21000-22000.csv', 
         'apartments_data_22000-24000.csv', 'apartments_data_24000-25000.csv',
         'apartments_data_25000-26000.csv', 'apartments_data_26000-28000.csv', 
         'apartments_data_28000-30000.csv', 'apartments_data_30000-to_end.csv']

dfs = [pd.read_csv(file) for file in files]
combined_df = pd.concat(dfs, axis=0, ignore_index=True)
combined_df

combined_df.to_csv('ap_data.csv', index=False)
